requirements.txt
aiogram
matplotlib
python-dotenv
aiosqlite
aiohttp
asyncpg
google-generativeai
